Time: 0 ms (100.00%), Space: 5.9 MB (55.27%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int reverse(long x) {
+    int reverse(long long x) {
         
         if(x<=9 && x>=0)return x;
